+++ src/LnaReaderCircular.cc
@@ -105,7 +105,7 @@ LnaReaderCircular::go_to(int frame)
 
     // Parse frame to the circular buffer
     for (int i = 0; i < m_num_models; i++) {
-      double tmp = (unsigned char)m_read_buffer[i + 1] / -24.0;
+      float tmp = (unsigned char)m_read_buffer[i + 1] / -24.0;
       m_log_prob_buffer[m_first_index] = tmp;
       m_first_index++;
       if (m_first_index >= m_log_prob_buffer.size())
